save matrix elements in the table
