+++ src/main/java/io/penguinstats/constant/Constant.java
@@ -2,15 +2,9 @@ package io.penguinstats.constant;
 
 public class Constant {
 
-	public static final Long[] ADD_TIME_POINTS =
-			new Long[] {0L, 1558989300000L, 1560045456000L, 1577145600000L, 1581105600000L};
-
 	public static final String API_V2 = "/api/v2";
 
 	public static class LastUpdateMapKeyName {
-		public static final String PAST_MATRIX_RESULT = "past_matrix_result";
-		public static final String CURRENT_MATRIX_RESULT = "current_matrix_result";
-		public static final String TREND_RESULT = "trend_result";
 		public static final String ITEM_LIST = "item_list";
 		public static final String ZONE_LIST = "zone_list";
 		public static final String STAGE_LIST = "stage_list";
