+++ src/main/java/io/penguinstats/enums/DropMatrixElementType.java
@@ -1,0 +1,14 @@
+package io.penguinstats.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum DropMatrixElementType {
+
+	REGULAR("reular"), TREND("trend");
+
+	private String name;
+
+}
