+++ src/main/java/io/penguinstats/model/query/GlobalTrendQuery.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 import io.penguinstats.enums.Server;
 import io.penguinstats.model.DropMatrixElement;
-import io.penguinstats.service.ItemDropService;
+import io.penguinstats.service.DropMatrixElementService;
 import io.penguinstats.util.QueryUtil;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -22,12 +22,12 @@ public class GlobalTrendQuery implements Serializable, BasicQuery {
 
 	private static final long serialVersionUID = 1L;
 
-	public GlobalTrendQuery(ItemDropService itemDropService) {
-		this.itemDropService = itemDropService;
+	public GlobalTrendQuery(DropMatrixElementService dropMatrixElementService) {
+		this.dropMatrixElementService = dropMatrixElementService;
 	}
 
 	@JsonIgnore
-	private ItemDropService itemDropService;
+	private DropMatrixElementService dropMatrixElementService;
 
 	private Server server;
 
@@ -40,7 +40,8 @@ public class GlobalTrendQuery implements Serializable, BasicQuery {
 	@Override
 	public List<DropMatrixElement> execute() throws Exception {
 		return QueryUtil.runQuery(
-				() -> itemDropService.generateSegmentedGlobalDropMatrixElements(server, interval, range), timeout);
+				() -> dropMatrixElementService.generateSegmentedGlobalDropMatrixElements(server, interval, range),
+				timeout);
 	}
 
 }
