+++ src/main/java/io/penguinstats/model/query/MatrixQuery.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 import io.penguinstats.enums.Server;
 import io.penguinstats.model.DropMatrixElement;
-import io.penguinstats.service.ItemDropService;
+import io.penguinstats.service.DropMatrixElementService;
 import io.penguinstats.util.QueryUtil;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -24,12 +24,12 @@ public class MatrixQuery implements Serializable, BasicQuery {
 
 	private static final long serialVersionUID = 1L;
 
-	public MatrixQuery(ItemDropService itemDropService) {
-		this.itemDropService = itemDropService;
+	public MatrixQuery(DropMatrixElementService dropMatrixElementService) {
+		this.dropMatrixElementService = dropMatrixElementService;
 	}
 
 	@JsonIgnore
-	private ItemDropService itemDropService;
+	private DropMatrixElementService dropMatrixElementService;
 
 	private Server server;
 
@@ -47,11 +47,10 @@ public class MatrixQuery implements Serializable, BasicQuery {
 
 	@Override
 	public List<DropMatrixElement> execute() throws Exception {
-		return QueryUtil
-				.runQuery(
-						() -> itemDropService.generateCustomDropMatrixElements(server, stageId, itemIds, start, end,
-								Optional.ofNullable(userID).map(userID -> Arrays.asList(userID)).orElse(null), null),
-						timeout);
+		return QueryUtil.runQuery(
+				() -> dropMatrixElementService.generateCustomDropMatrixElements(server, stageId, itemIds, start, end,
+						Optional.ofNullable(userID).map(userID -> Arrays.asList(userID)).orElse(null), null),
+				timeout);
 	}
 
 }
