+++ src/main/java/io/penguinstats/model/query/QueryFactory.java
@@ -1,26 +1,27 @@
 package io.penguinstats.model.query;
 
-import io.penguinstats.enums.QueryType;
-import io.penguinstats.service.ItemDropService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import io.penguinstats.enums.QueryType;
+import io.penguinstats.service.DropMatrixElementService;
 
 @Component("queryFactory")
 public class QueryFactory {
 
 	@Autowired
-	private ItemDropService itemDropService;
+	private DropMatrixElementService dropMatrixElementService;
 
 	public BasicQuery getQuery(QueryType type) throws Exception {
 		switch (type) {
 			case MATRIX:
-				return new MatrixQuery(itemDropService);
+				return new MatrixQuery(dropMatrixElementService);
 			case TREND:
-				return new TrendQuery(itemDropService);
+				return new TrendQuery(dropMatrixElementService);
 			case GLOBAL_MATRIX:
-				return new GlobalMatrixQuery(itemDropService);
+				return new GlobalMatrixQuery(dropMatrixElementService);
 			case GLOBAL_TREND:
-				return new GlobalTrendQuery(itemDropService);
+				return new GlobalTrendQuery(dropMatrixElementService);
 			default:
 				throw new Exception("Failed to create query for " + type);
 		}
