+++ src/main/java/io/penguinstats/task/UpdatePastDropMatrixTask.java
@@ -1,5 +1,6 @@
 package io.penguinstats.task;
 
+import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -7,8 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
+import io.penguinstats.enums.DropMatrixElementType;
 import io.penguinstats.enums.Server;
-import io.penguinstats.service.ItemDropService;
+import io.penguinstats.model.DropMatrixElement;
+import io.penguinstats.service.DropMatrixElementService;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
@@ -16,7 +19,7 @@ import lombok.extern.log4j.Log4j2;
 public class UpdatePastDropMatrixTask implements Task {
 
 	@Autowired
-	private ItemDropService itemDropService;
+	private DropMatrixElementService dropMatrixElementService;
 
 	@Scheduled(fixedRate = 43200000)
 	@Override
@@ -25,7 +28,12 @@ public class UpdatePastDropMatrixTask implements Task {
 
 		ExecutorService singleThreadExecutor = Executors.newSingleThreadExecutor();
 		for (Server server : Server.values()) {
-			singleThreadExecutor.execute(() -> itemDropService.refreshGlobalDropMatrixElements(server, true));
+			singleThreadExecutor.execute(() -> {
+				List<DropMatrixElement> elements =
+						dropMatrixElementService.generateGlobalDropMatrixElements(server, null, true);
+				dropMatrixElementService.batchDelete(DropMatrixElementType.REGULAR, server, true);
+				dropMatrixElementService.batchSave(elements);
+			});
 		}
 	}
